added components for styling inside docs/styling
