+++ src/app/docs/styling/page.mdx
@@ -1,4 +1,4 @@
-import Styling from '@/components/app/Styling'
+import Styling from './Styling'
 import Pagination from '@/components/app/Pagination'
 
 ## Styling
